DSND-2466: Refactored IntegrationDataGenerator to be an ArgumentsProvider
* IntegrationDataGenerator reads deltas directly from CHIPS
* No data read from CHIPS is written to the file system
* Refactored the parameterised test in ConsumerPositiveComprehensiveIT to take a feed from IntegrationDataGenerator
